Fixing float error

+++ artnet/ignite/train_model.py
@@ -212,7 +212,7 @@ def run(warmup_iterations=5000, batch_size=4, test_size=2000, epochs=10, log_int
             if 'iteration' in input_checkpoint:
                 trainer.state.iteration = input_checkpoint['iteration']
             else:
-                trainer.state.iteration = hparam_dict['training_set_size'] / batch_size * input_checkpoint['epoch']
+                trainer.state.iteration = int(hparam_dict['training_set_size'] / batch_size * input_checkpoint['epoch'])
 
 
     @trainer.on(Events.EPOCH_STARTED)
